+++ Tests/Runtime/TestScriptableObject.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace Tests.Runtime
+{
+	public class TestScriptableObject : ScriptableObject { }
+}